Move sudoku code into helpers.
